move kiosk creation to a mutator, fix state
